Rename dialog-related props and classes, make `WindowType` implementations customizable
